+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 25-MAR-2026 CEST
+Last change: 26-MAR-2026 CEST
 
 0.10.0
 ------
@@ -24,6 +24,7 @@ Last change: 25-MAR-2026 CEST
 - Only ``picasso.version.py`` determines software version globally, thus ``bumpversion`` is not needed anymore
 - Render GUI: more accessible saving/loading of FOVs as .txt files
 - Render GUI: keyboard combo for closing all localizations (Ctrl+Shift+Backspace or Ctrl+Shift+Delete)
+- Render GUI: legend is displayed on black background for better visibility
 
 *Bug fixes:*
 ++++++++++++